feat(aq1): L3.7 remove core SQLite test dependency
Move SQLite-backed nudge mode and dispatch coverage to integration tests using the shared runtime fixture, and record the AQ1 nudge-kind helper in the taxonomy inventory.
