+++ crates/atm-core/src/nudge_dispatch.rs
@@ -93,131 +93,3 @@ pub fn rebuild_received_hook_dispatch(
         nudge_mode,
     ))
 }
-
-#[cfg(test)]
-mod tests {
-    use std::fs;
-
-    use atm_storage_rusqlite::SqliteStorageBackend;
-
-    use super::*;
-    use crate::boundary::{PostSendBuiltInTarget, RosterEntry, RosterHarness, RosterMemberKind};
-    use crate::observability::NullObservability;
-    use crate::schema::AgentType;
-    use crate::send::{NudgeMode as SendNudgeMode, SendMessageSource, write_mail_with_runtime};
-    use crate::service_runtime::LocalFileNonClaudeOutbound;
-    use crate::types::{AgentName, ModelName, PaneId, TeamName};
-
-    fn roster_member(team: &TeamName, agent: &str, pane_id: Option<PaneId>) -> RosterEntry {
-        RosterEntry {
-            team_name: team.clone(),
-            agent_name: agent.parse().expect("agent"),
-            member_kind: RosterMemberKind::Permanent,
-            harness: RosterHarness::PythonGraft,
-            agent_type: AgentType::default(),
-            model: ModelName::default(),
-            recipient_pane_id: pane_id,
-            metadata_json: serde_json::Map::new(),
-        }
-    }
-
-    /// Builds a real SQLite-backed runtime (own on-disk database, not the
-    /// `atm-runtime-test-support` composition root, which returns a
-    /// `LocalServiceRuntime` from a second `atm_core` crate instance that
-    /// `#[cfg(test)]` code in this crate cannot name) with a seeded roster:
-    /// `sender` (no local backend) and `recipient` (tmux pane, so a write
-    /// triggers a `LocalSteer` dispatch this test can then rebuild).
-    fn setup() -> (tempfile::TempDir, LocalServiceRuntime, TeamName) {
-        let root = tempfile::tempdir().expect("temp root");
-        let database_path = root.path().join("mail.sqlite3");
-        let backend = SqliteStorageBackend::new(&database_path).expect("sqlite backend");
-        let runtime = LocalServiceRuntime::new_with_delivery_boundaries(
-            backend.message_store(),
-            backend.roster_store(),
-            backend.nudge_template_override_store(),
-            std::sync::Arc::new(LocalFileNonClaudeOutbound::new()),
-        )
-        .with_pending_nudge_store(backend.pending_nudge_store());
-        let team: TeamName = "test-team".parse().expect("team");
-        runtime
-            .roster_store
-            .save_roster(&atm_storage::RosterSnapshot {
-                team_name: team.clone(),
-                members: vec![
-                    roster_member(&team, "sender", None),
-                    roster_member(
-                        &team,
-                        "recipient",
-                        Some(PaneId::from_cli("%9").expect("pane")),
-                    ),
-                ],
-                refreshed_at: None,
-            })
-            .expect("seed roster");
-        (root, runtime, team)
-    }
-
-    fn write_request(
-        home_dir: &std::path::Path,
-        team: &TeamName,
-        nudge_mode: SendNudgeMode,
-    ) -> crate::send::WriteRequest {
-        crate::send::WriteRequest::new(
-            home_dir.to_path_buf(),
-            home_dir.to_path_buf(),
-            "sender".parse::<AgentName>().expect("sender"),
-            "recipient@test-team",
-            team.clone(),
-            SendMessageSource::Inline("rebuild me".to_owned()),
-            None,
-            false,
-            None,
-            false,
-        )
-        .expect("write request")
-        .with_nudge_mode(nudge_mode)
-    }
-
-    #[test]
-    fn rebuild_matches_write_time_dispatch_for_a_pending_tmux_recipient() {
-        let (root, runtime, team) = setup();
-        let home_dir = root.path().join("home");
-        fs::create_dir_all(&home_dir).expect("home dir");
-
-        let request = write_request(&home_dir, &team, SendNudgeMode::Immediate);
-        let outcome =
-            write_mail_with_runtime(request, &NullObservability, &runtime).expect("write succeeds");
-        let message_id = outcome.persisted_message_id();
-
-        let member = MemberKey::new(team, "recipient".parse().expect("agent"));
-        let rebuilt =
-            rebuild_received_hook_dispatch(&runtime, &member, message_id, NudgeKind::Queue)
-                .expect("rebuild succeeds")
-                .expect("dispatch rebuilt");
-
-        assert_eq!(rebuilt.kind, NudgeKind::Queue);
-        assert!(matches!(
-            rebuilt.target,
-            PostSendBuiltInTarget::LocalSteer(_)
-        ));
-        assert_eq!(rebuilt.event.message_id, message_id);
-    }
-
-    #[test]
-    fn rebuild_returns_none_for_a_message_not_addressed_to_member() {
-        let (root, runtime, team) = setup();
-        let home_dir = root.path().join("home");
-        fs::create_dir_all(&home_dir).expect("home dir");
-
-        let request = write_request(&home_dir, &team, SendNudgeMode::Immediate);
-        let outcome =
-            write_mail_with_runtime(request, &NullObservability, &runtime).expect("write succeeds");
-        let message_id = outcome.persisted_message_id();
-
-        let wrong_member = MemberKey::new(team, "sender".parse().expect("agent"));
-        let rebuilt =
-            rebuild_received_hook_dispatch(&runtime, &wrong_member, message_id, NudgeKind::Queue)
-                .expect("rebuild does not error");
-        assert!(rebuilt.is_none());
-    }
-}
